admin-blog/requirements.txt
beautifulsoup4
certifi
charset-normalizer
idna
numpy
pyfiglet
requests
soupsieve
typing_extensions
urllib3
websocket-client
gunicorn
flask-socketio
paho-mqtt
quart
hypercorn
proxychains
stem
python-socketio
sqlalchemy
aiosqlite
passlib
authlib
starlette
bcrypt
quart-session
redis
dotenv
aiohttp-socks
quart-cors
playwright
pysocks
requests[socks]
quart-authlib
apscheduler
websockets
reactivex